requirements.txt
Django>=3.2
djangorestframework>=3.12.4
channels>=3.0.3
channels-redis>=3.2.0